medical-recommender-for-woocommerce/requirements.txt
# Core dependencies for medical recommender system
pandas==2.1.4
numpy==1.24.4

# WooCommerce integration (Etap 2)
requests==2.31.0
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2

# Future dependencies (for next stages)
# Flask==3.0.0
# Flask-CORS==4.0.0 